refactor: 새로운 TokenAuthenticationInterceptor 사용하도록 리팩토링
