+++ src/main/java/nextstep/auth/authentication/token/TokenAuthenticationInterceptor.java
@@ -1,78 +1,42 @@
 package nextstep.auth.authentication.token;
 
+
 import com.fasterxml.jackson.databind.ObjectMapper;
-import nextstep.auth.authentication.AuthenticationException;
-import nextstep.auth.authentication.AuthenticationToken;
+import nextstep.auth.authentication.AuthenticationInterceptor;
 import nextstep.auth.authentication.UserDetailService;
 import nextstep.auth.context.Authentication;
 import nextstep.auth.token.JwtTokenProvider;
-import nextstep.auth.token.TokenRequest;
 import nextstep.auth.token.TokenResponse;
-import nextstep.member.domain.LoginMember;
 import org.springframework.http.MediaType;
-import org.springframework.web.servlet.HandlerInterceptor;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.stream.Collectors;
 
-public class TokenAuthenticationInterceptor implements HandlerInterceptor {
-    private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
+public class TokenAuthenticationInterceptor extends AuthenticationInterceptor {
 
-    private UserDetailService customUserDetailsService;
-    private JwtTokenProvider jwtTokenProvider;
+    private final ObjectMapper objectMapper;
+    private final JwtTokenProvider jwtTokenProvider;
 
-    public TokenAuthenticationInterceptor(UserDetailService customUserDetailsService, JwtTokenProvider jwtTokenProvider) {
-        this.customUserDetailsService = customUserDetailsService;
+    public TokenAuthenticationInterceptor(
+        UserDetailService customUserDetailsService,
+        TokenAuthenticationConverter tokenAuthenticationConverter,
+        JwtTokenProvider jwtTokenProvider,
+        ObjectMapper objectMapper
+    ) {
+        super(customUserDetailsService, tokenAuthenticationConverter);
+        this.objectMapper = objectMapper;
         this.jwtTokenProvider = jwtTokenProvider;
     }
 
     @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws IOException {
-        AuthenticationToken authenticationToken = convert(request);
-        Authentication authentication = authenticate(authenticationToken);
-
-        String payload = new ObjectMapper().writeValueAsString(authentication.getPrincipal());
+    public void afterAuthentication(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
+        String payload = objectMapper.writeValueAsString(authentication.getPrincipal());
         TokenResponse tokenResponse = new TokenResponse(jwtTokenProvider.createToken(payload));
 
-        String responseToClient = OBJECT_MAPPER.writeValueAsString(tokenResponse);
+        String responseToClient = objectMapper.writeValueAsString(tokenResponse);
         response.setStatus(HttpServletResponse.SC_OK);
         response.setContentType(MediaType.APPLICATION_JSON_VALUE);
         response.getOutputStream().print(responseToClient);
-
-        return false;
-    }
-
-    public AuthenticationToken convert(HttpServletRequest request) throws IOException {
-        TokenRequest tokenRequest = OBJECT_MAPPER.readValue(getRequestBody(request), TokenRequest.class);
-        String principal = tokenRequest.getEmail();
-        String credentials = tokenRequest.getPassword();
-
-        return new AuthenticationToken(principal, credentials);
-    }
-
-    private String getRequestBody(HttpServletRequest request) throws IOException {
-        return request.getReader()
-            .lines()
-            .collect(Collectors.joining(System.lineSeparator()));
-    }
-
-    public Authentication authenticate(AuthenticationToken authenticationToken) {
-        LoginMember loginMember = customUserDetailsService.loadUserByUsername(authenticationToken.getPrincipal());
-
-        checkAuthentication(loginMember, authenticationToken);
-
-        return new Authentication(loginMember);
-    }
-
-    private void checkAuthentication(LoginMember loginMember, AuthenticationToken authenticationToken) {
-        if (loginMember == null) {
-            throw new AuthenticationException();
-        }
-
-        if (!loginMember.checkPassword(authenticationToken.getCredentials())) {
-            throw new AuthenticationException();
-        }
     }
 }
